refactor(phase1d): DreamBooth/FineTuning/ControlNet データセットを分離
PR-1（dataset 系）の最終 PR。BaseDataset を継承する 3 つの具象データセット
クラスを、それぞれ独立したモジュールに移動。

## 新規モジュール

- library/dreambooth_dataset.py: DreamBoothDataset（約 379 行）
- library/finetuning_dataset.py: FineTuningDataset（約 280 行）
- library/controlnet_dataset.py: ControlNetDataset（約 279 行）

## 後方互換

train_util.py 側は約 760 行を削除し、3 クラスを re-export shim で維持:

    from library.dreambooth_dataset import DreamBoothDataset
    from library.finetuning_dataset import FineTuningDataset
    from library.controlnet_dataset import ControlNetDataset

## モジュール間依存

ControlNetDataset は内部で DreamBoothDataset / DreamBoothSubset を委譲先として
利用するため、library.controlnet_dataset → library.dreambooth_dataset の
依存がある。

## PR-1 シリーズ完了

これで train_util.py からのデータセット系分割（PR-1a〜1d）が完了。
- library/subset.py（PR-1a, #2333）
- library/caching.py（PR-1b, #2334）
- library/dataset.py（PR-1c, #2335）
- library/dreambooth_dataset.py / finetuning_dataset.py / controlnet_dataset.py
  （本 PR）

train_util.py は 6,888 行 → 3,774 行に縮小（約 45% 削減）。

## 動作確認

- ✅ 3 つの新規モジュール単独 import
- ✅ 全 PR-1 系列の identity（library.* と train_util re-export が同じクラス）
- ✅ 継承関係（DreamBoothDataset / FineTuningDataset / ControlNetDataset が
  BaseDataset のサブクラス）
- ✅ 全主要学習スクリプト（fine_tune, train_db, train_network, sdxl/sd3/flux/
  lumina/anima 系および各 control_net 系）の py_compile
