+++ library/train_util.py
@@ -138,764 +138,13 @@ from library.subset import BaseSubset, DreamBoothSubset, FineTuningSubset, Contr
 
 
 
-class DreamBoothDataset(BaseDataset):
-    IMAGE_INFO_CACHE_FILE = "metadata_cache.json"
+# DreamBooth / FineTuning / ControlNet datasets have moved to dedicated modules;
+# re-exported here for backward compatibility. New code should import from library.* directly.
+from library.dreambooth_dataset import DreamBoothDataset  # noqa: F401
+from library.finetuning_dataset import FineTuningDataset  # noqa: F401
+from library.controlnet_dataset import ControlNetDataset  # noqa: F401
 
-    # The is_training_dataset defines the type of dataset, training or validation
-    # if is_training_dataset is True -> training dataset
-    # if is_training_dataset is False -> validation dataset
-    def __init__(
-        self,
-        subsets: Sequence[DreamBoothSubset],
-        is_training_dataset: bool,
-        batch_size: int,
-        resolution,
-        network_multiplier: float,
-        enable_bucket: bool,
-        min_bucket_reso: int,
-        max_bucket_reso: int,
-        bucket_reso_steps: int,
-        bucket_no_upscale: bool,
-        prior_loss_weight: float,
-        train_inpainting: bool,
-        debug_dataset: bool,
-        validation_split: float,
-        validation_seed: Optional[int],
-        resize_interpolation: Optional[str],
-        skip_image_resolution: Optional[Tuple[int, int]] = None,
-    ) -> None:
-        super().__init__(
-            resolution,
-            network_multiplier,
-            train_inpainting,
-            debug_dataset,
-            resize_interpolation,
-            skip_image_resolution,
-        )
 
-        assert resolution is not None, f"resolution is required / resolution（解像度）指定は必須です"
-
-        self.batch_size = batch_size
-        self.size = min(self.width, self.height)  # 短いほう
-        self.prior_loss_weight = prior_loss_weight
-        self.latents_cache = None
-        self.is_training_dataset = is_training_dataset
-        self.validation_seed = validation_seed
-        self.validation_split = validation_split
-
-        self.enable_bucket = enable_bucket
-        if self.enable_bucket:
-            min_bucket_reso, max_bucket_reso = self.adjust_min_max_bucket_reso_by_steps(
-                resolution, min_bucket_reso, max_bucket_reso, bucket_reso_steps
-            )
-            self.min_bucket_reso = min_bucket_reso
-            self.max_bucket_reso = max_bucket_reso
-            self.bucket_reso_steps = bucket_reso_steps
-            self.bucket_no_upscale = bucket_no_upscale
-        else:
-            self.min_bucket_reso = None
-            self.max_bucket_reso = None
-            self.bucket_reso_steps = None  # この情報は使われない
-            self.bucket_no_upscale = False
-
-        def read_caption(img_path, caption_extension, enable_wildcard):
-            # captionの候補ファイル名を作る
-            base_name = os.path.splitext(img_path)[0]
-            base_name_face_det = base_name
-            tokens = base_name.split("_")
-            if len(tokens) >= 5:
-                base_name_face_det = "_".join(tokens[:-4])
-            cap_paths = [base_name + caption_extension, base_name_face_det + caption_extension]
-
-            caption = None
-            for cap_path in cap_paths:
-                if os.path.isfile(cap_path):
-                    with open(cap_path, "rt", encoding="utf-8") as f:
-                        try:
-                            lines = f.readlines()
-                        except UnicodeDecodeError as e:
-                            logger.error(f"illegal char in file (not UTF-8) / ファイルにUTF-8以外の文字があります: {cap_path}")
-                            raise e
-                        assert len(lines) > 0, f"caption file is empty / キャプションファイルが空です: {cap_path}"
-                        if enable_wildcard:
-                            caption = "\n".join([line.strip() for line in lines if line.strip() != ""])  # 空行を除く、改行で連結
-                        else:
-                            caption = lines[0].strip()
-                    break
-            return caption
-
-        def load_dreambooth_dir(subset: DreamBoothSubset):
-            if not os.path.isdir(subset.image_dir):
-                logger.warning(f"not directory: {subset.image_dir}")
-                return [], [], []
-
-            info_cache_file = os.path.join(subset.image_dir, self.IMAGE_INFO_CACHE_FILE)
-            use_cached_info_for_subset = subset.cache_info
-            if use_cached_info_for_subset:
-                logger.info(
-                    f"using cached image info for this subset / このサブセットで、キャッシュされた画像情報を使います: {info_cache_file}"
-                )
-                if not os.path.isfile(info_cache_file):
-                    logger.warning(
-                        f"image info file not found. You can ignore this warning if this is the first time to use this subset"
-                        + " / キャッシュファイルが見つかりませんでした。初回実行時はこの警告を無視してください: {metadata_file}"
-                    )
-                    use_cached_info_for_subset = False
-
-            if use_cached_info_for_subset:
-                # json: {`img_path`:{"caption": "caption...", "resolution": [width, height]}, ...}
-                with open(info_cache_file, "r", encoding="utf-8") as f:
-                    metas = json.load(f)
-                img_paths = list(metas.keys())
-                sizes: List[Optional[Tuple[int, int]]] = [meta["resolution"] for meta in metas.values()]
-
-                # we may need to check image size and existence of image files, but it takes time, so user should check it before training
-            else:
-                img_paths = glob_images(subset.image_dir, "*")
-                sizes: List[Optional[Tuple[int, int]]] = [None] * len(img_paths)
-
-                # new caching: get image size from cache files
-                strategy = LatentsCachingStrategy.get_strategy()
-                if strategy is not None:
-                    logger.info("get image size from name of cache files")
-
-                    # make image path to npz path mapping
-                    npz_paths = glob.glob(os.path.join(subset.image_dir, "*" + strategy.cache_suffix))
-                    npz_paths.sort(
-                        key=lambda item: item.rsplit("_", maxsplit=2)[0]
-                    )  # sort by name excluding resolution and cache_suffix
-                    npz_path_index = 0
-
-                    size_set_count = 0
-                    for i, img_path in enumerate(tqdm(img_paths)):
-                        l = len(os.path.splitext(img_path)[0])  # remove extension
-                        found = False
-                        while npz_path_index < len(npz_paths):  # until found or end of npz_paths
-                            # npz_paths are sorted, so if npz_path > img_path, img_path is not found
-                            if npz_paths[npz_path_index][:l] > img_path[:l]:
-                                break
-                            if npz_paths[npz_path_index][:l] == img_path[:l]:  # found
-                                found = True
-                                break
-                            npz_path_index += 1  # next npz_path
-
-                        if found:
-                            w, h = strategy.get_image_size_from_disk_cache_path(img_path, npz_paths[npz_path_index])
-                        else:
-                            w, h = None, None
-
-                        if w is not None and h is not None:
-                            sizes[i] = (w, h)
-                            size_set_count += 1
-                    logger.info(f"set image size from cache files: {size_set_count}/{len(img_paths)}")
-
-            if self.skip_image_resolution is not None:
-                filtered_img_paths = []
-                filtered_sizes = []
-                skip_image_area = self.skip_image_resolution[0] * self.skip_image_resolution[1]
-                for img_path, size in zip(img_paths, sizes):
-                    if size is None:  # no latents cache file, get image size by reading image file (slow)
-                        size = self.get_image_size(img_path)
-                    if size[0] * size[1] <= skip_image_area:
-                        continue
-                    filtered_img_paths.append(img_path)
-                    filtered_sizes.append(size)
-                if len(filtered_img_paths) < len(img_paths):
-                    logger.info(
-                        f"filtered {len(img_paths) - len(filtered_img_paths)} images by original resolution from {subset.image_dir}"
-                    )
-                img_paths = filtered_img_paths
-                sizes = filtered_sizes
-
-            # We want to create a training and validation split. This should be improved in the future
-            # to allow a clearer distinction between training and validation. This can be seen as a
-            # short-term solution to limit what is necessary to implement validation datasets
-            #
-            # We split the dataset for the subset based on if we are doing a validation split
-            # The self.is_training_dataset defines the type of dataset, training or validation
-            # if self.is_training_dataset is True -> training dataset
-            # if self.is_training_dataset is False -> validation dataset
-            if self.validation_split > 0.0:
-                # For regularization images we do not want to split this dataset.
-                if subset.is_reg is True:
-                    # Skip any validation dataset for regularization images
-                    if self.is_training_dataset is False:
-                        img_paths = []
-                        sizes = []
-                    # Otherwise the img_paths remain as original img_paths and no split
-                    # required for training images dataset of regularization images
-                else:
-                    img_paths, sizes = split_train_val(
-                        img_paths, sizes, self.is_training_dataset, self.validation_split, self.validation_seed
-                    )
-
-            logger.info(f"found directory {subset.image_dir} contains {len(img_paths)} image files")
-
-            if use_cached_info_for_subset:
-                captions = [metas[img_path]["caption"] for img_path in img_paths]
-                missing_captions = [img_path for img_path, caption in zip(img_paths, captions) if caption is None or caption == ""]
-            else:
-                # 画像ファイルごとにプロンプトを読み込み、もしあればそちらを使う
-                captions = []
-                missing_captions = []
-                for img_path in tqdm(img_paths, desc="read caption"):
-                    cap_for_img = read_caption(img_path, subset.caption_extension, subset.enable_wildcard)
-                    if cap_for_img is None and subset.class_tokens is None:
-                        logger.warning(
-                            f"neither caption file nor class tokens are found. use empty caption for {img_path} / キャプションファイルもclass tokenも見つかりませんでした。空のキャプションを使用します: {img_path}"
-                        )
-                        captions.append("")
-                        missing_captions.append(img_path)
-                    else:
-                        if cap_for_img is None:
-                            captions.append(subset.class_tokens)
-                            missing_captions.append(img_path)
-                        else:
-                            captions.append(cap_for_img)
-
-            self.set_tag_frequency(os.path.basename(subset.image_dir), captions)  # タグ頻度を記録
-
-            if missing_captions:
-                number_of_missing_captions = len(missing_captions)
-                number_of_missing_captions_to_show = 5
-                remaining_missing_captions = number_of_missing_captions - number_of_missing_captions_to_show
-
-                logger.warning(
-                    f"No caption file found for {number_of_missing_captions} images. Training will continue without captions for these images. If class token exists, it will be used. / {number_of_missing_captions}枚の画像にキャプションファイルが見つかりませんでした。これらの画像についてはキャプションなしで学習を続行します。class tokenが存在する場合はそれを使います。"
-                )
-                for i, missing_caption in enumerate(missing_captions):
-                    if i >= number_of_missing_captions_to_show:
-                        logger.warning(missing_caption + f"... and {remaining_missing_captions} more")
-                        break
-                    logger.warning(missing_caption)
-
-            if not use_cached_info_for_subset and subset.cache_info:
-                logger.info(f"cache image info for / 画像情報をキャッシュします : {info_cache_file}")
-                sizes = [self.get_image_size(img_path) for img_path in tqdm(img_paths, desc="get image size")]
-                matas = {}
-                for img_path, caption, size in zip(img_paths, captions, sizes):
-                    matas[img_path] = {"caption": caption, "resolution": list(size)}
-                with open(info_cache_file, "w", encoding="utf-8") as f:
-                    json.dump(matas, f, ensure_ascii=False, indent=2)
-                logger.info(f"cache image info done for / 画像情報を出力しました : {info_cache_file}")
-
-            # if sizes are not set, image size will be read in make_buckets
-            return img_paths, captions, sizes
-
-        logger.info("prepare images.")
-        num_train_images = 0
-        num_reg_images = 0
-        reg_infos: List[Tuple[ImageInfo, DreamBoothSubset]] = []
-        for subset in subsets:
-            num_repeats = subset.num_repeats if self.is_training_dataset else 1
-            if num_repeats < 1:
-                logger.warning(
-                    f"ignore subset with image_dir='{subset.image_dir}': num_repeats is less than 1 / num_repeatsが1を下回っているためサブセットを無視します: {num_repeats}"
-                )
-                continue
-
-            if subset in self.subsets:
-                logger.warning(
-                    f"ignore duplicated subset with image_dir='{subset.image_dir}': use the first one / 既にサブセットが登録されているため、重複した後発のサブセットを無視します"
-                )
-                continue
-
-            img_paths, captions, sizes = load_dreambooth_dir(subset)
-            if len(img_paths) < 1:
-                logger.warning(
-                    f"ignore subset with image_dir='{subset.image_dir}': no images found / 画像が見つからないためサブセットを無視します"
-                )
-                continue
-
-            if subset.is_reg:
-                num_reg_images += num_repeats * len(img_paths)
-            else:
-                num_train_images += num_repeats * len(img_paths)
-
-            for img_path, caption, size in zip(img_paths, captions, sizes):
-                info = ImageInfo(img_path, num_repeats, caption, subset.is_reg, img_path, subset.caption_dropout_rate)
-                info.resize_interpolation = (
-                    subset.resize_interpolation if subset.resize_interpolation is not None else self.resize_interpolation
-                )
-                if size is not None:
-                    info.image_size = size
-                if subset.is_reg:
-                    reg_infos.append((info, subset))
-                else:
-                    self.register_image(info, subset)
-
-            subset.img_count = len(img_paths)
-            self.subsets.append(subset)
-
-        images_split_name = "train" if self.is_training_dataset else "validation"
-        logger.info(f"{num_train_images} {images_split_name} images with repeats.")
-
-        self.num_train_images = num_train_images
-
-        logger.info(f"{num_reg_images} reg images with repeats.")
-        if num_train_images < num_reg_images:
-            logger.warning("some of reg images are not used / 正則化画像の数が多いので、一部使用されない正則化画像があります")
-
-        if num_reg_images == 0:
-            logger.warning("no regularization images / 正則化画像が見つかりませんでした")
-        else:
-            # num_repeatsを計算する：どうせ大した数ではないのでループで処理する
-            n = 0
-            first_loop = True
-            while n < num_train_images:
-                for info, subset in reg_infos:
-                    if first_loop:
-                        self.register_image(info, subset)
-                        n += info.num_repeats
-                    else:
-                        info.num_repeats += 1  # rewrite registered info
-                        n += 1
-                    if n >= num_train_images:
-                        break
-                first_loop = False
-
-        self.num_reg_images = num_reg_images
-
-
-class FineTuningDataset(BaseDataset):
-    def __init__(
-        self,
-        subsets: Sequence[FineTuningSubset],
-        batch_size: int,
-        resolution,
-        network_multiplier: float,
-        enable_bucket: bool,
-        min_bucket_reso: int,
-        max_bucket_reso: int,
-        bucket_reso_steps: int,
-        bucket_no_upscale: bool,
-        train_inpainting: bool,
-        debug_dataset: bool,
-        validation_seed: int,
-        validation_split: float,
-        resize_interpolation: Optional[str],
-        skip_image_resolution: Optional[Tuple[int, int]] = None,
-    ) -> None:
-        super().__init__(
-            resolution,
-            network_multiplier,
-            train_inpainting,
-            debug_dataset,
-            resize_interpolation,
-            skip_image_resolution,
-        )
-
-        self.batch_size = batch_size
-        self.size = min(self.width, self.height)  # 短いほう
-        self.latents_cache = None
-
-        self.enable_bucket = enable_bucket
-        if self.enable_bucket:
-            min_bucket_reso, max_bucket_reso = self.adjust_min_max_bucket_reso_by_steps(
-                resolution, min_bucket_reso, max_bucket_reso, bucket_reso_steps
-            )
-            self.min_bucket_reso = min_bucket_reso
-            self.max_bucket_reso = max_bucket_reso
-            self.bucket_reso_steps = bucket_reso_steps
-            self.bucket_no_upscale = bucket_no_upscale
-        else:
-            self.min_bucket_reso = None
-            self.max_bucket_reso = None
-            self.bucket_reso_steps = None  # この情報は使われない
-            self.bucket_no_upscale = False
-
-        self.num_train_images = 0
-        self.num_reg_images = 0
-
-        for subset in subsets:
-            if subset.num_repeats < 1:
-                logger.warning(
-                    f"ignore subset with metadata_file='{subset.metadata_file}': num_repeats is less than 1 / num_repeatsが1を下回っているためサブセットを無視します: {subset.num_repeats}"
-                )
-                continue
-
-            if subset in self.subsets:
-                logger.warning(
-                    f"ignore duplicated subset with metadata_file='{subset.metadata_file}': use the first one / 既にサブセットが登録されているため、重複した後発のサブセットを無視します"
-                )
-                continue
-
-            # メタデータを読み込む
-            if os.path.exists(subset.metadata_file):
-                if subset.metadata_file.endswith(".jsonl"):
-                    logger.info(f"loading existing JSOL metadata: {subset.metadata_file}")
-                    # optional JSONL format
-                    # {"image_path": "/path/to/image1.jpg", "caption": "A caption for image1", "image_size": [width, height]}
-                    metadata = {}
-                    with open(subset.metadata_file, "rt", encoding="utf-8") as f:
-                        for line in f:
-                            line_md = json.loads(line)
-                            image_md = {"caption": line_md.get("caption", "")}
-                            if "image_size" in line_md:
-                                image_md["image_size"] = line_md["image_size"]
-                            if "width" in line_md and "height" in line_md:
-                                image_md["image_size"] = [line_md["width"], line_md["height"]]
-                            if "tags" in line_md:
-                                image_md["tags"] = line_md["tags"]
-                            metadata[line_md["image_path"]] = image_md
-                else:
-                    # standard JSON format
-                    logger.info(f"loading existing metadata: {subset.metadata_file}")
-                    with open(subset.metadata_file, "rt", encoding="utf-8") as f:
-                        metadata = json.load(f)
-            else:
-                raise ValueError(f"no metadata / メタデータファイルがありません: {subset.metadata_file}")
-
-            if len(metadata) < 1:
-                logger.warning(
-                    f"ignore subset with '{subset.metadata_file}': no image entries found / 画像に関するデータが見つからないためサブセットを無視します"
-                )
-                continue
-
-            # Add full path for image
-            image_dirs = set()
-            if subset.image_dir is not None:
-                image_dirs.add(subset.image_dir)
-            for image_key in metadata.keys():
-                if not os.path.isabs(image_key):
-                    assert (
-                        subset.image_dir is not None
-                    ), f"image_dir is required when image paths are relative / 画像パスが相対パスの場合、image_dirの指定が必要です: {image_key}"
-                    abs_path = os.path.join(subset.image_dir, image_key)
-                else:
-                    abs_path = image_key
-                    image_dirs.add(os.path.dirname(abs_path))
-
-                # if image_key does not have extension, try to find image file with supported extensions
-                if not os.path.splitext(image_key)[1] or not os.path.exists(abs_path):  # no extension or file does not exist
-                    paths = glob_images(os.path.dirname(abs_path), os.path.basename(image_key))
-                    if len(paths) > 0:
-                        abs_path = paths[0]
-                    # If no file is found, we use *.npz file to get image size and for training
-
-                metadata[image_key]["abs_path"] = abs_path
-
-            # Enumerate existing npz files
-            strategy = LatentsCachingStrategy.get_strategy()
-            npz_paths = []
-            if strategy is not None:    # If `cache_latents` is not enabled (and no strategy is set), skip this part
-                for image_dir in image_dirs:
-                    npz_paths.extend(glob.glob(os.path.join(image_dir, "*" + strategy.cache_suffix)))
-            npz_paths = sorted(npz_paths, key=lambda item: len(os.path.basename(item)), reverse=True)  # longer paths first
-
-            # Match image filename longer to shorter because some images share same prefix
-            image_keys_sorted_by_length_desc = sorted(metadata.keys(), key=len, reverse=True)
-
-            # Collect tags and sizes
-            tags_list = []
-            size_set_from_metadata = 0
-            size_set_from_cache_filename = 0
-            num_filtered = 0
-            for image_key in image_keys_sorted_by_length_desc:
-                img_md = metadata[image_key]
-                caption = img_md.get("caption")
-                tags = img_md.get("tags")
-                image_size = img_md.get("image_size")
-                abs_path = img_md.get("abs_path")
-
-                # search npz if image_size is not given
-                npz_path = None
-                if image_size is None:
-                    image_without_ext = os.path.splitext(image_key)[0]
-                    for candidate in npz_paths:
-                        if candidate.startswith(image_without_ext):
-                            npz_path = candidate
-                            break
-                    if npz_path is not None:
-                        npz_paths.remove(npz_path)  # remove to avoid matching same file (share prefix)
-                        abs_path = npz_path
-
-                if caption is None:
-                    caption = ""
-
-                if subset.enable_wildcard:
-                    # tags must be single line (split by caption separator)
-                    if tags is not None:
-                        tags = tags.replace("\n", subset.caption_separator)
-
-                    # add tags to each line of caption
-                    if tags is not None:
-                        caption = "\n".join(
-                            [f"{line}{subset.caption_separator}{tags}" for line in caption.split("\n") if line.strip() != ""]
-                        )
-                        tags_list.append(tags)
-                else:
-                    # use as is
-                    if tags is not None and len(tags) > 0:
-                        if len(caption) > 0:
-                            caption = caption + subset.caption_separator
-                        caption = caption + tags
-                        tags_list.append(tags)
-
-                if caption is None:
-                    caption = ""
-
-                image_info = ImageInfo(image_key, subset.num_repeats, caption, False, abs_path, subset.caption_dropout_rate)
-                image_info.resize_interpolation = (
-                    subset.resize_interpolation if subset.resize_interpolation is not None else self.resize_interpolation
-                )
-
-                if image_size is not None:
-                    image_info.image_size = tuple(image_size)  # width, height
-                    size_set_from_metadata += 1
-                elif npz_path is not None:
-                    # get image size from npz filename
-                    w, h = strategy.get_image_size_from_disk_cache_path(abs_path, npz_path)
-                    image_info.image_size = (w, h)
-                    size_set_from_cache_filename += 1
-
-                if self.skip_image_resolution is not None:
-                    size = image_info.image_size
-                    if size is None:  # no image size in metadata or latents cache file, get image size by reading image file (slow)
-                        size = self.get_image_size(abs_path)
-                        image_info.image_size = size
-                    skip_image_area = self.skip_image_resolution[0] * self.skip_image_resolution[1]
-                    if size[0] * size[1] <= skip_image_area:
-                        num_filtered += 1
-                        continue
-
-                self.register_image(image_info, subset)
-
-            if size_set_from_cache_filename > 0:
-                logger.info(
-                    f"set image size from cache files: {size_set_from_cache_filename}/{len(image_keys_sorted_by_length_desc)}"
-                )
-            if size_set_from_metadata > 0:
-                logger.info(f"set image size from metadata: {size_set_from_metadata}/{len(image_keys_sorted_by_length_desc)}")
-            if num_filtered > 0:
-                logger.info(f"filtered {num_filtered} images by original resolution from {subset.metadata_file}")
-            self.num_train_images += len(metadata) * subset.num_repeats
-
-            # TODO do not record tag freq when no tag
-            self.set_tag_frequency(os.path.basename(subset.metadata_file), tags_list)
-            subset.img_count = len(metadata)
-            self.subsets.append(subset)
-
-
-class ControlNetDataset(BaseDataset):
-    def __init__(
-        self,
-        subsets: Sequence[ControlNetSubset],
-        batch_size: int,
-        resolution,
-        network_multiplier: float,
-        enable_bucket: bool,
-        min_bucket_reso: int,
-        max_bucket_reso: int,
-        bucket_reso_steps: int,
-        bucket_no_upscale: bool,
-        train_inpainting: bool,
-        debug_dataset: bool,
-        validation_split: float,
-        validation_seed: Optional[int],
-        resize_interpolation: Optional[str] = None,
-        skip_image_resolution: Optional[Tuple[int, int]] = None,
-    ) -> None:
-        super().__init__(
-            resolution,
-            network_multiplier,
-            train_inpainting,
-            debug_dataset,
-            resize_interpolation,
-            skip_image_resolution,
-        )
-
-        db_subsets = []
-        for subset in subsets:
-            assert (
-                not subset.random_crop
-            ), "random_crop is not supported in ControlNetDataset / random_cropはControlNetDatasetではサポートされていません"
-            db_subset = DreamBoothSubset(
-                subset.image_dir,
-                False,
-                None,
-                subset.caption_extension,
-                subset.cache_info,
-                False,
-                subset.num_repeats,
-                subset.shuffle_caption,
-                subset.caption_separator,
-                subset.keep_tokens,
-                subset.keep_tokens_separator,
-                subset.secondary_separator,
-                subset.enable_wildcard,
-                subset.color_aug,
-                subset.flip_aug,
-                subset.face_crop_aug_range,
-                subset.random_crop,
-                subset.caption_dropout_rate,
-                subset.caption_dropout_every_n_epochs,
-                subset.caption_tag_dropout_rate,
-                subset.caption_prefix,
-                subset.caption_suffix,
-                subset.token_warmup_min,
-                subset.token_warmup_step,
-                resize_interpolation=subset.resize_interpolation,
-            )
-            db_subsets.append(db_subset)
-
-        self.dreambooth_dataset_delegate = DreamBoothDataset(
-            db_subsets,
-            True,
-            batch_size,
-            resolution,
-            network_multiplier,
-            enable_bucket,
-            min_bucket_reso,
-            max_bucket_reso,
-            bucket_reso_steps,
-            bucket_no_upscale,
-            1.0,
-            train_inpainting,
-            debug_dataset,
-            validation_split,
-            validation_seed,
-            resize_interpolation,
-            skip_image_resolution,
-        )
-
-        # config_util等から参照される値をいれておく（若干微妙なのでなんとかしたい）
-        self.image_data = self.dreambooth_dataset_delegate.image_data
-        self.batch_size = batch_size
-        self.num_train_images = self.dreambooth_dataset_delegate.num_train_images
-        self.num_reg_images = self.dreambooth_dataset_delegate.num_reg_images
-        self.validation_split = validation_split
-        self.validation_seed = validation_seed
-        self.resize_interpolation = resize_interpolation
-
-        # assert all conditioning data exists
-        missing_imgs = []
-        cond_imgs_with_pair = set()
-        for image_key, info in self.dreambooth_dataset_delegate.image_data.items():
-            db_subset = self.dreambooth_dataset_delegate.image_to_subset[image_key]
-            subset = None
-            for s in subsets:
-                if s.image_dir == db_subset.image_dir:
-                    subset = s
-                    break
-            assert subset is not None, "internal error: subset not found"
-
-            if not os.path.isdir(subset.conditioning_data_dir):
-                logger.warning(f"not directory: {subset.conditioning_data_dir}")
-                continue
-
-            img_basename = os.path.splitext(os.path.basename(info.absolute_path))[0]
-            ctrl_img_path = glob_images(subset.conditioning_data_dir, img_basename)
-            if len(ctrl_img_path) < 1:
-                missing_imgs.append(img_basename)
-                continue
-            ctrl_img_path = ctrl_img_path[0]
-            ctrl_img_path = os.path.abspath(ctrl_img_path)  # normalize path
-
-            info.cond_img_path = ctrl_img_path
-            cond_imgs_with_pair.add(os.path.splitext(ctrl_img_path)[0])  # remove extension because Windows is case insensitive
-
-        extra_imgs = []
-        for subset in subsets:
-            conditioning_img_paths = glob_images(subset.conditioning_data_dir, "*")
-            conditioning_img_paths = [os.path.abspath(p) for p in conditioning_img_paths]  # normalize path
-            extra_imgs.extend([p for p in conditioning_img_paths if os.path.splitext(p)[0] not in cond_imgs_with_pair])
-
-        assert (
-            len(missing_imgs) == 0
-        ), f"missing conditioning data for {len(missing_imgs)} images / 制御用画像が見つかりませんでした: {missing_imgs}"
-        if len(extra_imgs) > 0:
-            logger.warning(f"extra conditioning data for {len(extra_imgs)} images / 余分な制御用画像があります: {extra_imgs}")
-
-        self.conditioning_image_transforms = IMAGE_TRANSFORMS
-
-    def set_current_strategies(self):
-        return self.dreambooth_dataset_delegate.set_current_strategies()
-
-    def make_buckets(self):
-        self.dreambooth_dataset_delegate.make_buckets()
-        self.bucket_manager = self.dreambooth_dataset_delegate.bucket_manager
-        self.buckets_indices = self.dreambooth_dataset_delegate.buckets_indices
-
-    def cache_latents(self, vae, vae_batch_size=1, cache_to_disk=False, is_main_process=True):
-        return self.dreambooth_dataset_delegate.cache_latents(vae, vae_batch_size, cache_to_disk, is_main_process)
-
-    def new_cache_latents(self, model: Any, accelerator: Accelerator):
-        return self.dreambooth_dataset_delegate.new_cache_latents(model, accelerator)
-
-    def new_cache_text_encoder_outputs(self, models: List[Any], is_main_process: bool):
-        return self.dreambooth_dataset_delegate.new_cache_text_encoder_outputs(models, is_main_process)
-
-    def __len__(self):
-        return self.dreambooth_dataset_delegate.__len__()
-
-    def __getitem__(self, index):
-        example = self.dreambooth_dataset_delegate[index]
-
-        bucket = self.dreambooth_dataset_delegate.bucket_manager.buckets[
-            self.dreambooth_dataset_delegate.buckets_indices[index].bucket_index
-        ]
-        bucket_batch_size = self.dreambooth_dataset_delegate.buckets_indices[index].bucket_batch_size
-        image_index = self.dreambooth_dataset_delegate.buckets_indices[index].batch_index * bucket_batch_size
-
-        conditioning_images = []
-
-        for i, image_key in enumerate(bucket[image_index : image_index + bucket_batch_size]):
-            image_info = self.dreambooth_dataset_delegate.image_data[image_key]
-
-            target_size_hw = example["target_sizes_hw"][i]
-            original_size_hw = example["original_sizes_hw"][i]
-            crop_top_left = example["crop_top_lefts"][i]
-            flipped = example["flippeds"][i]
-            cond_img = load_image(image_info.cond_img_path)
-
-            if self.dreambooth_dataset_delegate.enable_bucket:
-                assert (
-                    cond_img.shape[0] == original_size_hw[0] and cond_img.shape[1] == original_size_hw[1]
-                ), f"size of conditioning image is not match / 画像サイズが合いません: {image_info.absolute_path}"
-
-                cond_img = resize_image(
-                    cond_img,
-                    original_size_hw[1],
-                    original_size_hw[0],
-                    target_size_hw[1],
-                    target_size_hw[0],
-                    self.resize_interpolation,
-                )
-
-                # TODO support random crop
-                # 現在サポートしているcropはrandomではなく中央のみ
-                h, w = target_size_hw
-                ct = (cond_img.shape[0] - h) // 2
-                cl = (cond_img.shape[1] - w) // 2
-                cond_img = cond_img[ct : ct + h, cl : cl + w]
-            else:
-                # assert (
-                #     cond_img.shape[0] == self.height and cond_img.shape[1] == self.width
-                # ), f"image size is small / 画像サイズが小さいようです: {image_info.absolute_path}"
-                # resize to target
-                if cond_img.shape[0] != target_size_hw[0] or cond_img.shape[1] != target_size_hw[1]:
-                    cond_img = resize_image(
-                        cond_img,
-                        cond_img.shape[0],
-                        cond_img.shape[1],
-                        target_size_hw[1],
-                        target_size_hw[0],
-                        self.resize_interpolation,
-                    )
-
-            if flipped:
-                cond_img = cond_img[:, ::-1, :].copy()  # copy to avoid negative stride
-
-            cond_img = self.conditioning_image_transforms(cond_img)
-            conditioning_images.append(cond_img)
-
-        example["conditioning_images"] = torch.stack(conditioning_images).to(memory_format=torch.contiguous_format).float()
-
-        return example
-
-
-# behave as Dataset mock
 # Caching functions have moved to library.caching; re-exported here for backward compatibility.
 # New code should import from library.caching directly.
 from library.caching import (  # noqa: F401
